requirements.txt
fastapi
uvicorn
python-dotenv
prometheus-fastapi-instrumentator
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-classic
langchain-chroma
sentence-transformers
chromadb
docx2txt
pypdf
